+++ xarray/backends/h5netcdf_.py
@@ -71,14 +71,25 @@ class H5NetCDFStore(WritableCFDataStore):
     """
 
     def __init__(
-        self, filename, mode="r", format=None, group=None, lock=None, autoclose=False
+        self,
+        filename,
+        mode="r",
+        format=None,
+        group=None,
+        lock=None,
+        autoclose=False,
+        invalid_netcdf=None,
     ):
         import h5netcdf
 
         if format not in [None, "NETCDF4"]:
             raise ValueError("invalid format for h5netcdf backend")
 
-        self._manager = CachingFileManager(h5netcdf.File, filename, mode=mode)
+        kwargs = {"invalid_netcdf": invalid_netcdf}
+
+        self._manager = CachingFileManager(
+            h5netcdf.File, filename, mode=mode, kwargs=kwargs
+        )
 
         if lock is None:
             if mode == "r":